Initial release
After an intense week or two, here is the first release of BatchMDL, a Blender addon for batch exporting static props to Source Engine.
I tried to make a workflow with as few repeated actions as possible, so that you can quickly export dozens of models with minimal setup.
It should work on Linux, with Wine if there is no native executable, but I haven't tested this yet.