Add support for SMD export
This is necessary for older games.
I reused code from the SourceOps SMD exporter.
SMD is default for compatibility, but FBX is recommended when possible, because the exporter is faster.
I'm not super happy with the aesthetics of making that use_collision temp variable, but line length was too long.
Got rid of translate=False for game properties because those generic names should translate fine.
Dunno how that affects descriptions though.